+++ source/refined-github.ts
@@ -185,7 +185,7 @@ import './features/keyboard-navigation';
 import './features/vertical-front-matter';
 import './features/use-first-commit-message-for-new-prs';
 import './features/linkify-user-edit-history-popup';
-import './features/cleanup-repo-filelist-actions';
+import './features/clean-repo-filelist-actions';
 import './features/prevent-duplicate-pr-submission';
 import './features/remove-label-faster';
 import './features/clean-conversation-headers';
